requirements.txt
asteval==0.9.32
asttokens==2.4.1
attrs==23.2.0
bokeh==3.4.1
broadbean==0.14.0
cf_xarray==0.9.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dask==2024.5.1
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
executing==2.0.1
fonttools==4.51.0
fsspec==2024.5.0
future==1.0.0
h5netcdf==1.3.0
h5py==3.11.0
importlib-metadata==7.0.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.4
jsonref==1.1.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lmfit==1.3.1
locket==1.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
opentelemetry-api==1.24.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
partd==1.4.2
pillow==10.3.0
platformdirs==4.2.2
plottr @ git+https://github.com/qipe-nlab/plottr.git@1728baeb5dfaf1befcf359c761f17708a33b9764
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.1.0
pyelftools==0.31
Pygments==2.18.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
pytz==2024.1
PyVISA==1.14.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
qcodes==0.45.0
QtPy==2.4.1
referencing==0.35.1
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
schema==0.7.7
scipy==1.13.1
six==1.16.0
stack-data==0.6.3
tabulate==0.9.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
uncertainties==3.1.7
versioningit==3.1.1
watchdog==4.0.0
wcwidth==0.2.13
websockets==12.0
widgetsnbextension==4.0.10
wrapt==1.16.0
xarray==2024.5.0
xhistogram==0.3.2
xyzservices==2024.4.0
zhinst-core==24.4.57831
zhinst-qcodes==0.5.3
zhinst-timing-models==24.4.57831
zhinst-toolkit==0.6.3
zhinst-utils==0.4.1
zipp==3.18.2
